% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_metrics.R
\name{aug_metrics}
\alias{aug_metrics}
\title{Augmented Data Evaluation Metrics}
\usage{
aug_metrics(df)
}
\arguments{
\item{df}{A dataframe object. This should be a structured dataset where each
column represents a variable and each row represents an observation.}
}
\description{
Calculate metrics for data to be used in augmentation.
}
