% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_random_missingness.R
\name{inject_na}
\alias{inject_na}
\title{Inject Random Missingness}
\usage{
inject_na(data = ., prop = 0.1, vars = NULL, method)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{prop}{A numeric value indicating what proportion of the data should be
injected with missingness.}

\item{vars}{An optional vector of variable names that should be injected with
missingness. If \code{vars} is not specified, the function will default to
injecting missingness across the entire dataset.}

\item{method}{A character value indicating whether to determine random
proportion of missingness by column or row. Options include \code{"row"} to
inject missingness by row or \code{"column"} to inject missingness by
column.}
}
\description{
Inject Random Missingness
}
