% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_random_noise.R
\name{inject_noise}
\alias{inject_noise}
\title{Inject Random Noise}
\usage{
inject_noise(data = ., iter = 1, vars = NULL, noise_value = 1, digits = NA)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{iter}{A numeric value indicating how many iterations of noise injection
should be applied. The default is \code{1}.}

\item{vars}{An optional vector of variable names that should be injected with
random noise. If \code{vars} is not specified, the function will default to
injecting random noise across the entire dataset.}

\item{noise_value}{A numeric value indicating the absolute range of noise
from \code{0} to \code{X} to be added or subtracted as noise. The default
is \code{1}.}

\item{digits}{A numeric value indicating the number of digits to round to.}
}
\description{
Inject Random Noise
}
