% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf_analysis}
\alias{turf_analysis}
\title{Total Unduplicated Reach & Frequency Analysis}
\usage{
turf_analysis(data, vars = NULL)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{vars}{An optional vector of variable names that should be included in
the TURF analysis. If \code{vars} is not specified, the function will
default to running TURF across the entire dataset.}
}
\description{
Total Unduplicated Reach & Frequency Analysis
}
