% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf_optimize}
\alias{turf_optimize}
\title{Total Unduplicated Reach & Frequency Optimizer}
\usage{
turf_optimize(data, vars = NULL, combos = c(2:3))
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{vars}{An optional vector of variable names that should be included in
the TURF analysis. If \code{vars} is not specified, the function will
default to running TURF across the entire dataset.}

\item{combos}{A vector of numeric values indicating the range of feature
combinations to include in the TURF analysis. The default is \code{c(2:3)}.}
}
\description{
Total Unduplicated Reach & Frequency Optimizer
}
